Struct air_trace_handler::TraceHandler

pub struct TraceHandler { /* private fields */ }

Implementations

impl TraceHandler

pub fn from_trace( prev_trace: ExecutionTrace, current_trace: ExecutionTrace, ) -> Self

pub fn trace_pos(&self) -> Result<TracePos, IntConversionError>

Returns size of elements inside result trace and intended to provide a position of next inserted elements.

pub fn into_result_trace(self) -> ExecutionTrace

pub fn as_result_trace(&self) -> &ExecutionTrace

pub fn subgraph_sizes(&self) -> (TraceLen, TraceLen)

pub fn update_generation( &mut self, trace_pos: TracePos, generation: GenerationIdx, ) -> Result<(), GenerationCompactificationError>

impl TraceHandler

pub fn meet_call_start(&mut self) -> TraceHandlerResult<MergerCallResult>

Should be called at the beginning of a call execution.

pub fn meet_call_end(&mut self, call_result: CallResult)

Should be called when a call instruction was executed successfully. It adds the supplied state to the result trace.

impl TraceHandler

pub fn meet_ap_start(&mut self) -> TraceHandlerResult<MergerApResult>

pub fn meet_ap_end(&mut self, ap_result: ApResult)

impl TraceHandler

pub fn meet_canon_start(&mut self) -> TraceHandlerResult<MergerCanonResult>

pub fn meet_canon_end(&mut self, canon_result: CanonResult)

impl TraceHandler

pub fn meet_par_start(&mut self) -> TraceHandlerResult<()>

pub fn meet_par_subgraph_end( &mut self, subgraph_type: SubgraphType, ) -> TraceHandlerResult<()>

impl TraceHandler

pub fn meet_fold_start(&mut self, fold_id: u32) -> TraceHandlerResult<()>

pub fn meet_iteration_start( &mut self, fold_id: u32, value_pos: TracePos, ) -> TraceHandlerResult<()>

pub fn meet_iteration_end(&mut self, fold_id: u32) -> TraceHandlerResult<()>

pub fn meet_back_iterator(&mut self, fold_id: u32) -> TraceHandlerResult<()>

pub fn meet_generation_end(&mut self, fold_id: u32) -> TraceHandlerResult<()>

pub fn meet_fold_end(&mut self, fold_id: u32) -> TraceHandlerResult<()>